chore(config): move config out of the packaged `constants` directory
